chore: 여행 성향 테스트 수정 로직 이동 (#131)
